Renamed Functions to viewMeFunctions
